requirements.txt
asgiref==3.5.2
gunicorn==20.1.0
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.2.0
Django==4.1.3
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-filter==22.1
django-heroku==0.3.1
django-import-export==3.0.1
django-tinymce==3.5.0
djangorestframework==3.14.0
Pillow==9.3.0
